Add new files
An updated README.md will be added with more organization and a better description. A `/docs` folder will be added for when the website is made and published. a `/Designs` folder was added for each device size so that when the coding process is started developers have a visual to work off of. The background image in the designs were added into the repo.